python_source/requirements.txt
rsa
boto3
botocore
google-api-core
google-cloud-datastore 
google-cloud-secret-manager
google-cloud-logging
google-cloud-pubsub
google-cloud-storage
mysql-connector-python